+++ qbit/admin/src/main/java/io/advantageous/qbit/metrics/support/LocalStatsCollector.java
@@ -32,7 +32,7 @@ public class LocalStatsCollector implements StatReplicator, QueueCallBackHandler
     final static class Metric {
 
         final StatList stats;
-        int value;
+        long value;
         final String name;
         final MetricType type;
 
@@ -65,7 +65,7 @@ public class LocalStatsCollector implements StatReplicator, QueueCallBackHandler
 
 
     @Override
-    public void replicateCount(String name, int count, long time) {
+    public void replicateCount(String name, long count, long time) {
 
         Metric metric = statsMap.get(name);
         if (metric == null) {
@@ -79,7 +79,7 @@ public class LocalStatsCollector implements StatReplicator, QueueCallBackHandler
     }
 
     @Override
-    public void replicateLevel(String name, int level, long time) {
+    public void replicateLevel(String name, long level, long time) {
 
 
         Metric metric = statsMap.get(name);
@@ -95,7 +95,7 @@ public class LocalStatsCollector implements StatReplicator, QueueCallBackHandler
     }
 
     @Override
-    public void replicateTiming(String name, int timing, long time) {
+    public void replicateTiming(String name, long timing, long time) {
 
         Metric metric = statsMap.get(name);
         if (metric == null) {
